selection sort completed
Selection sort algorithm starts by compairing first two elements of an
array and swapping if necessary, i.e., if you want to sort the elements
of array in ascending order and if the first element is greater than
second then, you need to swap the elements but, if the first element is
smaller than second, leave the elements as it is. Then, again first
element and third element are compared and swapped if necessary. This
process goes on until first and last element of an array is compared.
This completes the first step of selection sort.

If there are n elements to be sorted then, the process mentioned above
should be repeated n-1 times to get required result. But, for better
performance, in second step, comparison starts from second element
because after first step, the required number is automatically placed at
the first (i.e, In case of sorting in ascending order, smallest element
will be at first and in case of sorting in descending order, largest
element will be at first.). Similarly, in third step, comparison starts
from third element and so on.